+++ internal/config/config_test.go
@@ -4,6 +4,8 @@ import (
 	"net/http"
 	"net/url"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 func TestOverrides(t *testing.T) {
@@ -42,4 +44,31 @@ func TestOverrides(t *testing.T) {
 		t.Fail()
 	}
 
+}
+
+func TestReplaceVariables(t *testing.T) {
+
+	// Set a list of secrets to use
+	CfgData.Secrets = map[string]string{
+		"TEST_SECRET": "123456",
+		"HOST":        "host123",
+	}
+
+	cfgFile := []byte(`some config
+	test config: hello
+	test config host: $HOST
+	test config secret: $TEST_SECRET
+	test config no secret: TEST_SECRET
+	testboth: $TEST_SECRET $HOST`)
+
+	replaceVariables(&cfgFile)
+
+	assert.Equal(t, `some config
+	test config: hello
+	test config host: host123
+	test config secret: 123456
+	test config no secret: TEST_SECRET
+	testboth: 123456 host123`,
+		string(cfgFile),
+		"Config variable replacements should match")
 }